Create training_testing_assessing
This script trains a logit model with elastic net for predicting the probability of a client renewing its contract 1 year after ending its relationship with the organization. It takes into consideration many types of variables, ranging from sociodemographic to financing or economic. The variables are not given as they are, many of them are calculated, the ones that use local money values (colombian peso) are updated to present values multiplying by inflation, ages are updated, information about the local market is inputted through features as the average rate of interest of the month, and a long etc.

Also, as the actual sample with which I had to work was heavily imbalanced toward negative predictions, I used SMOTE for upsampling the positive values (clients that renewed) and Tomek-Links for keeping the more extrinsically different type of observations between both types of predictions.

For SMOTE'ing' I had to chosse a number of clusters from which to start 'brewing' new artificial observations, I designed a series of for loop to test five different combinations of clusters and, also, for choosing the best elastic net hyperparameter, I designed a for loop that test 10 different values for alpha ranging from 0 (ridge) to 1.0 (lasso). In total, we end with 55 different fits.

I could explain a lot more processes in detail, but I think the script speaks for itself, this is a very complex but complete piece of code, of which I am very proud of. Suffice to say that the trained model had a very high F1 value (0.77) and I am sure I can continue improving it.